Pavilion

+++ sistema/includes/ajax.php
@@ -3101,7 +3101,7 @@ if ($_POST['action'] == 'EditarAdeudo') {
         $descuento = isset($_POST['descuento']) ? floatval($_POST['descuento']) : 0.0;
         $semanas_totales = intval($_POST['semanas_totales']);
         $fecha_final = mysqli_real_escape_string($conection, $_POST['fecha_final']);
-        $total_abonado = intval($_POST['total_abonado']);
+        $total_abonado = floatval($_POST['total_abonado']);
 
         // Ejecutar la consulta
         $query_insertar = mysqli_query(
